Use xmake always.
Note that if the user has already configured with ninja then this
won't overwrite that. I found when testing that if the path has
() in it then this will fail when using xcommon cmake and ninja, but not
with Makefiles. Windows puts brackets in directory names when you
download something with a duplicate name, e.g. sandbox(1). This
will save time helping users but power users can still use ninja