Update reverb max room size and sample rate config parameters.
Change them to constants as updating them at run time is not supported.

Also update the reverb stage test to test the passing case, this will
ensure future regressions get caught.

Also add the constants property to the Stage class so that pipeline.py
does not need to access a private member.